Add upload scenario (mobile --> server) with JPEGmini compression on server
