+++ app/build.gradle
@@ -40,4 +40,9 @@ dependencies {
     compile 'com.koushikdutta.async:androidasync:2.1.7'
     compile 'com.afollestad.material-dialogs:core:0.8.5.9'
     compile 'com.android.support:support-v4:23.4.0'
+
+
+    compile 'com.squareup.okhttp3:logging-interceptor:3.3.1'
+
+
 }
