+++ app/src/main/res/layout/content_main.xml
@@ -67,4 +67,13 @@
         android:layout_centerHorizontal="true"
         android:layout_marginBottom="50dp" />
 
+    <Button
+        android:id="@+id/button_upload_batch"
+        android:text="Batch Upload PB"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentLeft="true"
+        android:layout_marginBottom="50dp" />
+
 </RelativeLayout>
